refactor(model): minimize MiniMax M3 model diffs
